api: add `metrics.py`
Move `Metrics` class implementation to separate file
to reduce number of lines in `api.main` module and to
keep the various modules more organized.
Otherwise pylint will complain with the below error:
"api/main.py: C0302: Too many lines in module (1044/1000) (too-many-lines)"
